refactor: erd 변경에 맞춰 코드 수정
Entity, Service, Repository, DTO 등 ERD 변경에 따라 생기는 오류 및 바뀌는 로직 수정

+++ src/main/java/com/example/cafe/domain/PopularMenu.java
@@ -1,7 +1,7 @@
 package com.example.cafe.domain;
 
 public interface PopularMenu {
-    Integer getMenu();
-    Integer getTotal();
+    Long getMenu();
+    Long getTotal();
 
 }
