+++ src/main/java/com/example/cafe/dto/reponsedto/PointResponseDto.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class PointResponseDto {
 
-    private Integer memberId;
-    private Integer point;
+    private Long memberId;
+    private Long point;
 }
